# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreatePromoRequest:
    properties:
      code:
        maxLength: 50
        minLength: 3
        type: string
      description:
        type: string
      discount:
        maximum: 100
        minimum: 1
        type: number
      end_date:
        type: string
      max_usage:
        type: integer
      start_date:
        type: string
    required:
    - code
    - discount
    - end_date
    - start_date
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.MidtransNotificationRequest:
    properties:
      fraud_status:
        type: string
      gross_amount:
        type: string
      order_id:
        type: string
      payment_type:
        type: string
      settlement_time:
        type: string
      signature_key:
        type: string
      status_code:
        type: string
      transaction_id:
        type: string
      transaction_status:
        type: string
    required:
    - gross_amount
    - order_id
    - signature_key
    - status_code
    - transaction_status
    type: object
  dto.RegisterRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - confirm_password
    - email
    - name
    - password
    type: object
  dto.UpdatePromoRequest:
    properties:
      description:
        type: string
      discount:
        maximum: 100
        minimum: 1
        type: number
      end_date:
        type: string
      is_active:
        type: boolean
      max_usage:
        type: integer
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@cinemticket.com
    name: Cinema Ticketing Support
  description: REST API untuk sistem manajemen tiket bioskop. Mendukung manajemen
    studio, film, jadwal tayang, kursi, booking tiket, transaksi, promo diskon, dan
    laporan penjualan.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cinema Ticketing API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentikasi user dan mendapatkan JWT token.
      parameters:
      - description: Kredensial login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login berhasil, token tersedia
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Email atau password salah
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun pengguna baru dengan role 'user'.
      parameters:
      - description: Data registrasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User berhasil didaftarkan
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
      summary: Register user baru
      tags:
      - Auth
  /payment/notification:
    post:
      consumes:
      - application/json
      description: Receives a Midtrans webhook and verifies its status directly with
        Midtrans.
      parameters:
      - description: Midtrans notification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MidtransNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Midtrans payment notification
      tags:
      - Payment
  /promo:
    get:
      description: Admin dapat melihat semua kode promo yang tersedia.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List semua promo
      tags:
      - Promo
    post:
      consumes:
      - application/json
      description: Admin dapat membuat kode promo diskon dengan persentase tertentu
        dan periode berlaku.
      parameters:
      - description: Data promo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePromoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat promo baru
      tags:
      - Promo
  /promo/{id}:
    delete:
      description: Admin dapat menghapus kode promo yang sudah tidak digunakan.
      parameters:
      - description: Promo ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus promo
      tags:
      - Promo
    get:
      description: Mengambil detail satu kode promo berdasarkan ID-nya.
      parameters:
      - description: Promo ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detail promo
      tags:
      - Promo
    put:
      consumes:
      - application/json
      description: Admin dapat memperbarui data promo yang sudah ada.
      parameters:
      - description: Promo ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data promo yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePromoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update promo
      tags:
      - Promo
  /promo/validate/{code}:
    get:
      description: User dapat mengecek apakah kode promo valid dan melihat preview
        harga setelah diskon sebelum melakukan booking.
      parameters:
      - description: Kode promo
        in: path
        name: code
        required: true
        type: string
      - description: Harga total sebelum diskon
        in: query
        name: total_price
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Validasi kode promo
      tags:
      - Promo
...
  /transaction/{transaction_id}/pay:
    post:
      description: Membuat Snap transaction dan mengembalikan URL pembayaran Midtrans.
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat pembayaran Midtrans
      tags:
      - Payment
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"